refactor: NetworkManager의 내용을 Service와 Util로 분리
